Fix accordion panel height transitions
The uneven movement at the start and end of accordion panel height transitions is caused by the element content height which can never be negative as would be the case when the panel height is set to 0 but some padding is defined. An additional wrapper div is required to get around this.