Add wrapper element around drawer and backdrop
As is the case with dialogs, nesting the drawer inside the backdrop element leads to opacity changes on the latter also being applied to the drawer. A drawer should not fade when animated so it would be best to define it as a separate item alongside the backdrop inside a wrapper.