Allow collapsibles to be shown by default
Collapsibles are initially hidden and can be shown when a trigger is activated. Some components may require that they are shown by default. Items can transition only from a known height so one has to be explicitly set for these elements.

Using requestAnimationFrame() is necessary although exactly how it helps in these circumstances is poorly understood.