Use :last-of-type to select last segment
The `:last-of-type` CSS pseudo-class is slightly better suited for styling the last accordion segment because `:last-child` might not yield the desired results if some other element is placed last inside the accordion.

Note that `:last-of-type` may also be unreliable as it refers to the last element, not class, of the designated type.