Fix premature leave transition end
Using `{ once: true }` ensures an event listener is invoked at most once after being added. This can be problematic as `transitionend` may be triggered because of a transition happening on a descendant element. Checking the event target against the element to which the event handler has been attached ensures class removal happens only when transitions on the chosen element have completed.